Removing dependency `byteorder`.
