+++ crates/svm-codec/Cargo.toml
@@ -20,8 +20,6 @@ svm-layout = { path = "../svm-layout" }
 svm-abi-encoder = { path = "../svm-abi/encoder" }
 svm-abi-decoder = { path = "../svm-abi/decoder" }
 svm-sdk = { path = "../svm-sdk" }
-byteorder = "1.3.2"
-bit-vec = "0.6.1"
 serde_json = "1.0"
 
 [dev-dependencies]
