+++ crates/svm-codec/src/serialize.rs
